refactor: extract ACP backend types, errors, and utils into separate modules
Extract from backend.ts (1,306 → 1,030 lines):
- acp-backend-types.ts (240 lines): All interface/type definitions
- acp-backend-errors.ts (19 lines): AcpBackendLifecycleError, AcpBackendRuntimeUnavailableError
- acp-backend-utils.ts (135 lines): Pure utility functions (scopeFromInput, attachmentFromResult, readPackageVersion, etc.)

- TSC clean, madge clean (0 circular deps)
- All types re-exported from backend.ts for backward compatibility
- No behavioral changes — pure structural refactoring

Part of Architecture Fix Plan Priority 2: reduce file sizes toward 500-line quality gate.
